Bootstrap applied; added about.html; migrated game app to .js file; added css
A bunch of big changes. Substantial prettification of game app page.
Added a bunch of CSS/HTML prettiness through Bootstrap plugin. Used a
javascript/bootstrap calculator example to put in a pretty keypad.
(Still need to figure out why the button outlines aren’t showing up.)

Migrated the readme.md page into an “about.html” page that incorporates
a lot of the app design and learning methodology.

The actual game is no longer working, as it isn’t displaying anything
in the calculator field. That’s in large part due to the fact that the
calculator field doesn’t exist, and needs to be written.